requirements.txt

streamlit
openai==1.66.3
git+https://github.com/openai/whisper.git
python-pptx
ffmpeg-python
